+++ go/server/license_service/license_service.go
@@ -40,7 +40,7 @@ type CreateLicenseParams struct {
 
 // CreateProductParams defines model for CreateProductParams.
 type CreateProductParams struct {
-	Amount           *int                    `json:"amount"`
+	Amount           *int64                  `json:"amount"`
 	Currency         *string                 `json:"currency"`
 	Disabled         *bool                   `json:"disabled"`
 	EntitlementsJson *map[string]interface{} `json:"entitlements_json"`
@@ -94,7 +94,7 @@ type FindLicenseParams struct {
 
 // FindProductParams defines model for FindProductParams.
 type FindProductParams struct {
-	Amount           *int                    `json:"amount,omitempty"`
+	Amount           *int64                  `json:"amount,omitempty"`
 	CreatedAt        *time.Time              `json:"created_at,omitempty"`
 	Currency         *string                 `json:"currency,omitempty"`
 	Disabled         *bool                   `json:"disabled,omitempty"`
@@ -153,7 +153,7 @@ type PatchLicenseParams struct {
 
 // PatchProductParams defines model for PatchProductParams.
 type PatchProductParams struct {
-	Amount           *int                    `json:"amount"`
+	Amount           *int64                  `json:"amount"`
 	Currency         *string                 `json:"currency"`
 	Disabled         *bool                   `json:"disabled,omitempty"`
 	EntitlementsJson *map[string]interface{} `json:"entitlements_json,omitempty"`
@@ -214,7 +214,7 @@ type ProcessIOSPurchaseRequest struct {
 // Product defines model for Product.
 type Product struct {
 	// The currency amount in cents ($19.99)
-	Amount    *int       `json:"amount,omitempty"`
+	Amount    *int64     `json:"amount,omitempty"`
 	CreatedAt *time.Time `json:"created_at,omitempty"`
 
 	// The currency code (ISO 4217)
@@ -262,7 +262,7 @@ type UpdateLicenseParams struct {
 
 // UpdateProductParams defines model for UpdateProductParams.
 type UpdateProductParams struct {
-	Amount           *int                    `json:"amount"`
+	Amount           *int64                  `json:"amount"`
 	Currency         *string                 `json:"currency"`
 	Disabled         *bool                   `json:"disabled"`
 	EntitlementsJson *map[string]interface{} `json:"entitlements_json"`
